requirements.txt
openvino>=2024.4.0
openvino-genai>=2024.4.0
soundcard>=0.4.3
scipy>=1.11
numpy>=1.24
keyboard>=0.13.5
pyperclip>=1.8.2
webrtcvad-wheels>=2.0.14
noisereduce>=3.0
PySide6>=6.7